Fix : placeCardCarousel made to component
