+++ frontend/src/components/map/ourPlaceMap/ourPlaceMap.module.css
@@ -1,4 +1,4 @@
 .map {
-  width: 100vw;
-  height: 100vh;
+  width: 100%;
+  height: 100%;
 }
