+++ frontend/src/components/placeSlider/placeSlider.module.css
@@ -1,0 +1,6 @@
+.placeCardContainer {
+  height: 17vh;
+  width: 90%;
+  margin: auto;
+  background-color: white;
+}
